+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds 'focus' and 'datatype' builtin hooks.
 - The 'unzip' hook now supports tar and gz.
 - Post-hooks in fetchez.core was ignoring entry changes, this fixes that.
+- We now use yaml files for config, including presets
 
 ## [0.4.2] - 2026-02-21
 ### Added